restructured file directory to have datacheck_meta and checks split up within data_checks folder
