+++ datachecks/__init__.py
@@ -1,22 +1,22 @@
 """General datachecks directory."""
 
-from datachecks.data_checks.data_check import DataCheck
-from datachecks.data_checks.data_check_message_code import DataCheckMessageCode
-from datachecks.data_checks.data_check_action import DataCheckAction
-from datachecks.data_checks.data_check_action_option import (
+from datachecks.data_checks.checks.data_check import DataCheck
+from datachecks.data_checks.datacheck_meta.data_check_message_code import DataCheckMessageCode
+from datachecks.data_checks.datacheck_meta.data_check_action import DataCheckAction
+from datachecks.data_checks.datacheck_meta.data_check_action_option import (
     DataCheckActionOption,
     DCAOParameterType,
     DCAOParameterAllowedValuesType,
 )
-from datachecks.data_checks.data_check_action_code import DataCheckActionCode
-from datachecks.data_checks.data_checks import DataChecks
-from datachecks.data_checks.data_check_message import (
+from datachecks.data_checks.datacheck_meta.data_check_action_code import DataCheckActionCode
+from datachecks.data_checks.checks.data_checks import DataChecks
+from datachecks.data_checks.datacheck_meta.data_check_message import (
     DataCheckMessage,
     DataCheckWarning,
     DataCheckError,
 )
-from datachecks.data_checks.data_check_message_type import DataCheckMessageType
-from datachecks.data_checks.id_columns_data_check import IDColumnsDataCheck
+from datachecks.data_checks.datacheck_meta.data_check_message_type import DataCheckMessageType
+from datachecks.data_checks.checks.id_columns_data_check import IDColumnsDataCheck
 
 from datachecks.problem_types.problem_types import ProblemTypes
 from datachecks.problem_types.utils import (
