+++ datachecks/data_checks/datacheck_meta/data_check_action_option.py
@@ -1,8 +1,8 @@
 """Recommended action returned by a DataCheck."""
 from enum import Enum
 
-from datachecks.data_checks.data_check_action import DataCheckAction
-from datachecks.data_checks.data_check_action_code import DataCheckActionCode
+from datachecks.data_checks.datacheck_meta.data_check_action import DataCheckAction
+from datachecks.data_checks.datacheck_meta.data_check_action_code import DataCheckActionCode
 from datachecks.utils import classproperty
